refactor(consolidate_to_lotto): Code consolidation to Lotto class
- moving lottary functionality to Lotto class
- making use of companion objects
